Move imports connection logic to own module
Move imports connection logic to own module

- port connection logic in own module
- shared imports-doc related stuff now in index

Rename browser-history module as imports

- all the non-UI/background imports-related stuff lives in here
- bookmarks planned to be added in `import-bookmarks` submodule, as with `import-history`
- update the Codetour line as well as background entrypoint import
